fix type of length arg to xmlrpc_read_base64_old()
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@374 98333e67-4a24-44d7-a75c-e53540dd3050

+++ trunk/xmlrpc-c/src/xmlrpc_data.c
@@ -539,7 +539,7 @@ xmlrpc_read_base64(xmlrpc_env *           const envP,
 void
 xmlrpc_read_base64_old(xmlrpc_env *           const envP,
                        const xmlrpc_value *   const valueP,
-                       unsigned int *         const lengthP,
+                       size_t *               const lengthP,
                        const unsigned char ** const byteStringValueP) {
 
     validateType(envP, valueP, XMLRPC_TYPE_BASE64);
